Remove `img` from `LepTess` struct
It was only there so it could be manually dropped when Tess gets
dropped.

This is no longer needed because:
img (Pix) now gets automatically dropped by Rust
img does not need to live at least as long as tess, because `set_image`
makes its own copy of the image.