Deal with Windows test failures
First if failed to compile because a constant (capi::L_CLONE) was of
type i32 rather than u32 as expected. Fortunately, it then gets cast at
i32, so i32 works out better for us.

I've changed the function `get_box` to expect an i32 rather than a u32,
and coerce `L_CLONE` to be i32.

I would have made `get_box` generic, but I couldn't see how to do
generic coercions.

A test that set the environment variable TESSDATA_PREFIX wasn't working.
As this isn't a core piece of functionality, I've set it not to run on
Windows.

Finally a doc test was relying on tesseract having a default location to
get trained data. This wasn't working as vcpkg tesseract doesn't have a
default location. As the test wasn't about not having default locations,
I set a default location.

https://github.com/houqp/leptess/issues/3