Fix CI by upgrading Rust version
The CI is currently failing for `cargo clippy`

https://app.circleci.com/pipelines/github/houqp/leptess/100/workflows/cb968766-5459-44f4-ba7b-9195214b9cf4/jobs/125

For things that aren't being picked up locally.

I think features it is flagging as experimental have since been merged.
For example:

```
error[E0658]: non exhaustive is an experimental feature (see issue #44109)
  --> /usr/local/cargo/registry/src/github.com-1ecc6299db9ec823/libloading-0.6.7/src/error.rs:23:1
   |
23 | #[non_exhaustive]
   | ^^^^^^^^^^^^^^^^^
```

https://github.com/rust-lang/rust/issues/44109

As such, I think it's time to upgrade the rust version of the CI.