Commit variables.rs
This will make it easier to see what effect changes to the build script
will have.

I have also updated the build script to python.

Why not use build.rs?
Output isn't commitable, making it hard to see how changes to the build
script affect it.

Why not use Rust?
It's easier to use an interpreted language from a makefile.

This new generator is nicer in that it has a clean parse and output
seperation. If the old one were to be used again, it would be good to
get that seperation.

This new one is also nicer in that it's only focussed on the variables
file. build.rs is allowed to build many files, so it can't be as
focussed even when only one output file is desired.