Variables: Move from_bytes_with_nul up a level
This simplifies the inside of the match expression by not repeating the
same code for every match arm.

I think it would be ok to use `from_bytes_with_nul_unchecked` here as we
know all the strings end in \0 and it's easy to check none contain \0
except at the end.

But ultimately, what would the advantage be?

https://doc.rust-lang.org/std/ffi/struct.CStr.html#method.from_bytes_with_nul_unchecked